Restart processes less aggressively
Insert a 1 second pause after each stop. Never stop respawning essentially, but thanks to the sleep we shouldn't create too much load if a process continues crashing.

Comes from a problem on boot in which Elasticsearch is not ready to
serve requests, but the processes try to connect to it to get the
index-metadata from the index used as a key-value store.